master_id and node_id are long vars
master_id and node_id are long vars

+++ server/include/server.h
@@ -71,7 +71,7 @@ typedef struct server {
 	long		node_id;	/**< Node id, server_id for M/S or local_index for Galera */
 	int		rlag;		/**< Replication Lag for Master / Slave replication */
 	unsigned long	node_ts;	/**< Last timestamp set from M/S monitor module */
-	int		master_id;	/**< Master server id of this node */
+	long		master_id;	/**< Master server id of this node */
 	int		depth;		/**< Replication level in the tree */
 	long		*slaves;	/**< Slaves of this node */
 } SERVER;
